Expand docblock of JoinSub to allow other models
The previous version of this docblock used the template `TModelClass`.
However, this template was already in use at the class level,
effectively limiting the argument to `joinSub` to be a QueryBuilder of
the same type.

The test proved that this was allowed:

```
User::query()->joinSub(
    User::query()->whereIn('id', [1, 2, 3]),
    'users',
    'users.id',
    'users.id'
);
```

However, in the wild, I noticed that
`User::query()->joinSub(Post::query(), ...)` was _not_ allowed.

This change fixes that, simply by renaming the template in the `joinSub`
docblock.